updated model usage method
