+++ flagai/auto_model/auto_loader.py
@@ -4,9 +4,9 @@
 import importlib
 import os
 import copy
-from flagai.model.file_utils import _get_model_id
+from flagai.model.file_utils import _get_model_id, _get_checkpoint_path, _get_vocab_path, _get_model_files
+from flagai.model.aquila2.modeling_aquila import AquilaForCausalLM
 import torch
-
 
 class LazyImport(object):
 
@@ -207,8 +207,54 @@ class AutoLoader:
             print(f"All supported models are {list(MODEL_DICT.keys())}")
             return
         if task_name == "aquila2":
-            from flagai.model.aquila2_model import Aquila2Model
-            from accelerate import init_empty_weights, load_checkpoint_and_dispatch
+            download_path = os.path.join(model_dir, model_name)
+            
+            if not os.path.exists(download_path):
+                # Try to download from ModelHub
+                try:
+                    model_id = _get_model_id(model_name)
+                except:
+                    raise FileNotFoundError("Model name not found in local path and ModelHub")
+                if model_id and model_id != "null":
+                    model_files = eval(_get_model_files(model_name))
+                    print("model files:" + str(model_files))
+                    for file_name in model_files:
+                        if not file_name.endswith("bin"):
+                            _get_vocab_path(download_path, file_name, model_id)
+
+                    if os.path.exists(
+                            os.path.join(download_path, 'config.json')):
+                        if os.getenv('ENV_TYPE') == 'deepspeed+mpu':
+                            model_parallel_size = int(os.getenv("MODEL_PARALLEL_SIZE"))
+                            if model_parallel_size > 1:
+                                # if gpus == nums_of_modelhub_models
+                                # can load
+                                # else need to download the pytorch_model.bin and to recut.
+                                model_hub_parallel_size = 0
+                                for f in model_files:
+                                    if "pytorch_model_" in f:
+                                        model_hub_parallel_size += 1
+                        else:
+                            model_parallel_size = 1
+
+                        if "pytorch_model_01.bin" in model_files and model_parallel_size > 1 and model_hub_parallel_size == model_parallel_size:
+                            # Only to download the model slices(megatron-lm).
+                            for file_to_load in model_files:
+                                if "pytorch_model_" in file_to_load:
+                                    _get_checkpoint_path(download_path, file_to_load,
+                                                        model_id)
+
+                        elif 'pytorch_model.bin' in model_files:
+                            checkpoint_path = _get_checkpoint_path(
+                                download_path, 'pytorch_model.bin', model_id)
+                        else:
+                            checkpoint_merge = {}
+                            # maybe multi weights files
+                            for file_to_load in model_files:
+                                if "pytorch_model-0" in file_to_load:
+                                    _get_checkpoint_path(download_path, file_to_load,
+                                                        model_id)            
+
             if qlora_dir:
                 from transformers import BitsAndBytesConfig
                 quantization_config=BitsAndBytesConfig(
@@ -217,11 +263,14 @@ class AutoLoader:
                     bnb_4bit_quant_type="nf4",
                     bnb_4bit_compute_dtype=torch_dtype,
                 )
-            model = Aquila2Model.from_pretrain(model_dir, model_name,
+
+
+            model = AquilaForCausalLM.from_pretrained(download_path,
                                                     low_cpu_mem_usage=low_cpu_mem_usage, torch_dtype=torch_dtype,
                                                     quantization_config=quantization_config)
             
             model.eval()
+            # from accelerate import load_checkpoint_and_dispatch
             # model = load_checkpoint_and_dispatch(
             #                 model, model_dir+model_name, device_map="balanced", no_split_module_classes=["LlamaDecoderLayer"])
             if not qlora_dir:
@@ -236,13 +285,9 @@ class AutoLoader:
                 print("Qlora modules loaded")
             from transformers import AutoTokenizer
             tokenizer = AutoTokenizer.from_pretrained(model_dir+model_name)
-            #args.cuda_index = 0
-            # device = f"cuda"
             self.model = model 
             self.tokenizer = tokenizer 
-
         else:
-
             brief_model_name = MODEL_DICT[model_name][2]
             model_type = MODEL_DICT[model_name][3]
             # The dir to save config, vocab and model.
